+++ app/src/main/java/com/ubademy_mobile/activities/SplashActivity.kt
@@ -14,6 +14,7 @@ class SplashActivity : AppCompatActivity() {
 
         Handler(Looper.getMainLooper()).postDelayed({
             startActivity(Intent(this,LoginActivity::class.java))
+            finish()
         }, 2000)
     }
 }
