+++ app/src/main/java/com/ubademy_mobile/services/data/Curso.kt
@@ -1,0 +1,11 @@
+package com.ubademy_mobile.services
+
+import java.util.*
+
+data class Curso(val id: String?,
+                 val id_creador: String?,
+                 val titulo: String?,
+                 val descripcion: String?,
+                 val estado: String?,
+                 val fecha_creacion: String?,
+                 val fecha_actualizacion: String?)
